feat: [LT-1141] Add metrics.
This took some doing because the API they provide is the stupidest,
most overengineered piece of crap I've ever seen. And it's not even
really threadsafe! (Technically it is because the data is accumulated in
a thread-safe manner, but it's super context-sensitive so it's only good
if the various threads _write different metrics_. I don't know what those
guys are smoking exactly)

I'd started to write a thingy that emulated this whole mess but the
additional depencencies aren't so bad, so I figured in the end
that it wasn't a good idea, even though what they're using is super
overkill.