feat: [LT-1141] log too large messages (#14)
* feat: [LT-1141] WIP: log (and DLQ) events that are too long for mParticle.

Currently the size limit is fixed to 256 000 bytes (about 256 kb - since
we're not sure if they're using metric or base 2 kb we err on the side of
caution). This is roughly enough for 650 messages per batch give or take
a few.

* refactor: [LT-1141] Add more tests and reorganize code a bit to make that easier.

1. Get rid of the explicit arguments constructor - instead, pass a map (which is actually
   `System.getenv()` by default) so we can test that the env vars get parsed properly.
2. More test coverage
3. Made an attempt at replacing all assertEquals etc. with assertThat.

* refactor: [LT-1141] Stricter visibility to avoid unneeded exposure of a setter.

* feat: [LT-1141] Add metrics.

This took some doing because the API they provide is the stupidest,
most overengineered piece of crap I've ever seen. And it's not even
really threadsafe! (Technically it is because the data is accumulated in
a thread-safe manner, but it's super context-sensitive so it's only good
if the various threads _write different metrics_. I don't know what those
guys are smoking exactly)

I'd started to write a thingy that emulated this whole mess but the
additional depencencies aren't so bad, so I figured in the end
that it wasn't a good idea, even though what they're using is super
overkill.

* fmt: [LT-1141] grrr checkstyle

* test: [LT-1141] Improve test coverage.

1. Add mock test that verifies we properly send the metric out on message too large
2. Add some logic to verify that our metrics implementation sends out metrics as expected
   to the "sink", properly formatted

* fmt: [LT-1141] checkstyle goes brrr